# Use chunks in TMap's folds and transform