… benchmarks/src/main/scala/zio/stm/TMapOpsBenchmarks.scala
… core-tests/shared/src/test/scala/zio/stm/TArraySpec.scala
… core-tests/shared/src/test/scala/zio/stm/TMapSpec.scala
… core/shared/src/main/scala/zio/stm/ZSTM.scala