### Summary

- Added `toChunk` to `TMap` and `TArray`.
- Removed `indices` from `TArray` transformations.
- Added `fold` / `foldM` benchmarks.
- Replaced lists with chunks inside folds and transformations.
- Improve efficiency and safety of chunked foreach (ZSTM).

### Results

#### Folds

Before:

```
Benchmark                (size)   Mode  Cnt       Score      Error  Units
TMapOpsBenchmarks.fold        0  thrpt   15  258658.256 ± 8415.195  ops/s
TMapOpsBenchmarks.fold       10  thrpt   15  184811.477 ± 1966.108  ops/s
TMapOpsBenchmarks.fold      100  thrpt   15   22520.429 ±   76.867  ops/s
TMapOpsBenchmarks.fold     1000  thrpt   15    2396.229 ±    6.618  ops/s
TMapOpsBenchmarks.fold    10000  thrpt   15     109.687 ±    1.259  ops/s
TMapOpsBenchmarks.fold   100000  thrpt   15       6.887 ±    0.156  ops/s
TMapOpsBenchmarks.foldM       0  thrpt   15  281237.383 ± 9172.810  ops/s
TMapOpsBenchmarks.foldM      10  thrpt   15  165526.805 ± 2169.245  ops/s
TMapOpsBenchmarks.foldM     100  thrpt   15   19590.089 ±  120.782  ops/s
TMapOpsBenchmarks.foldM    1000  thrpt   15    1929.585 ±    5.396  ops/s
TMapOpsBenchmarks.foldM   10000  thrpt   15     100.054 ±    3.493  ops/s
TMapOpsBenchmarks.foldM  100000  thrpt   15       5.975 ±    0.102  ops/s
```

After:

```
Benchmark                (size)   Mode  Cnt       Score      Error  Units
TMapOpsBenchmarks.fold        0  thrpt   15  229489.328 ± 7150.990  ops/s
TMapOpsBenchmarks.fold       10  thrpt   15  207447.701 ± 4881.882  ops/s
TMapOpsBenchmarks.fold      100  thrpt   15   28134.487 ±   82.870  ops/s
TMapOpsBenchmarks.fold     1000  thrpt   15    2652.985 ±   16.872  ops/s
TMapOpsBenchmarks.fold    10000  thrpt   15     122.824 ±    0.596  ops/s
TMapOpsBenchmarks.fold   100000  thrpt   15       7.985 ±    0.066  ops/s
TMapOpsBenchmarks.foldM       0  thrpt   15  241354.670 ± 7260.260  ops/s
TMapOpsBenchmarks.foldM      10  thrpt   15  186563.730 ± 4521.997  ops/s
TMapOpsBenchmarks.foldM     100  thrpt   15   24254.184 ±  114.819  ops/s
TMapOpsBenchmarks.foldM    1000  thrpt   15    1931.638 ±   10.842  ops/s
TMapOpsBenchmarks.foldM   10000  thrpt   15     104.820 ±    0.640  ops/s
TMapOpsBenchmarks.foldM  100000  thrpt   15       6.992 ±    0.205  ops/s
```

#### Transformations

Before:

```
TMapOpsBenchmarks.transform        0  thrpt   15  148985.179 ± 3818.584  ops/s
TMapOpsBenchmarks.transform       10  thrpt   15  118974.349 ± 2685.904  ops/s
TMapOpsBenchmarks.transform      100  thrpt   15   15679.842 ±   42.883  ops/s
TMapOpsBenchmarks.transform     1000  thrpt   15    1839.234 ±   22.481  ops/s
TMapOpsBenchmarks.transform    10000  thrpt   15      96.660 ±    0.992  ops/s
TMapOpsBenchmarks.transform   100000  thrpt   15       7.093 ±    0.112  ops/s
TMapOpsBenchmarks.transformM       0  thrpt   15  144693.581 ± 4076.923  ops/s
TMapOpsBenchmarks.transformM      10  thrpt   15   97583.697 ± 1395.881  ops/s
TMapOpsBenchmarks.transformM     100  thrpt   15   11675.240 ±   85.219  ops/s
TMapOpsBenchmarks.transformM    1000  thrpt   15    1237.373 ±    3.750  ops/s
TMapOpsBenchmarks.transformM   10000  thrpt   15      74.736 ±    1.810  ops/s
TMapOpsBenchmarks.transformM  100000  thrpt   15       5.120 ±    0.080  ops/s
```

After:

```
Benchmark                     (size)   Mode  Cnt       Score      Error  Units
TMapOpsBenchmarks.transform        0  thrpt   15  144259.175 ± 5290.334  ops/s
TMapOpsBenchmarks.transform       10  thrpt   15  145538.922 ± 3996.517  ops/s
TMapOpsBenchmarks.transform      100  thrpt   15   20636.735 ±  311.573  ops/s
TMapOpsBenchmarks.transform     1000  thrpt   15    2371.484 ±  120.327  ops/s
TMapOpsBenchmarks.transform    10000  thrpt   15     120.811 ±    4.753  ops/s
TMapOpsBenchmarks.transform   100000  thrpt   15       7.330 ±    0.329  ops/s
TMapOpsBenchmarks.transformM       0  thrpt   15  139583.672 ± 3749.977  ops/s
TMapOpsBenchmarks.transformM      10  thrpt   15  105385.824 ± 1371.753  ops/s
TMapOpsBenchmarks.transformM     100  thrpt   15   12748.558 ±   37.286  ops/s
TMapOpsBenchmarks.transformM    1000  thrpt   15    1385.189 ±    4.494  ops/s
TMapOpsBenchmarks.transformM   10000  thrpt   15      63.870 ±    0.910  ops/s
TMapOpsBenchmarks.transformM  100000  thrpt   15       4.902 ±    0.077  ops/s
```