@adamgfraser 2020-04-13
Now that `Chunk` extends `IndexedSeq` there may be some additional simplifications we can do here. Also `ChunkBuilder` could be useful if you need to build an array and then convert it into a `Chunk`.