@mijicd 2020-04-13
@jdegoes @adamgfraser Yes, `ChunkBuilder` was missing before, I'm updating the PR now. Just integrating changes from master revealed some performance drop in `transformM`, so I'll see what I can do there.